requirements.txt
langchain
langgraph
google-genai
sentence-transformers
faiss-cpu
python-dotenv
